requirements.txt
Flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
pandas==2.1.0
flask-session==0.5.0
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
gunicorn==21.2.0
